+++ src/gui/gui-canvas.js
@@ -314,8 +314,11 @@ export function DisplayCanvas() {
 
   function getStyleKey(style) {
     return (style.strokeWidth > 0 ? style.strokeColor + '~' + style.strokeWidth +
-      '~' + (style.lineDash ? style.lineDash + '~' : '') : '') +
-      (style.fillColor || '') + (style.opacity < 1 ? '~' + style.opacity : '');
+      '~' + (style.lineDash ? style.lineDash + '~' : '') +
+      (style.strokeOpacity >= 0 ? style.strokeOpacity + '~' : '') : '') +
+      (style.fillColor || '') +
+      (style.fillOpacity ? '~' + style.fillOpacity : '') +
+      (style.opacity < 1 ? '~' + style.opacity : '');
   }
 
   return _self;
@@ -464,9 +467,9 @@ function getPathStart(ext, lineScale) {
   return function(ctx, style) {
     var strokeWidth;
     ctx.beginPath();
-    if (style.opacity >= 0) {
-      ctx.globalAlpha = style.opacity;
-    }
+    // if (style.opacity >= 0) {
+    //   ctx.globalAlpha = style.opacity;
+    // }
     if (style.strokeWidth > 0) {
       strokeWidth = style.strokeWidth;
       if (pixRatio > 1) {
@@ -489,14 +492,22 @@ function getPathStart(ext, lineScale) {
 }
 
 function endPath(ctx, style) {
-  if (style.fillColor) ctx.fill();
+  var fo = style.opacity >= 0 ? style.opacity : 1,
+      so = fo;
+  if (style.strokeOpacity >= 0) so *= style.strokeOpacity;
+  if (style.fillOpacity >= 0) fo *= style.fillOpacity;
+  if (style.fillColor) {
+    ctx.globalAlpha = fo;
+    ctx.fill();
+  }
   if (style.strokeWidth > 0) {
+    ctx.globalAlpha = so;
     ctx.stroke();
     if (style.lineDash) {
       ctx.lineCap = 'round';
       ctx.setLineDash([]);
     }
   }
-  if (style.opacity >= 0) ctx.globalAlpha = 1;
+  ctx.globalAlpha = 1;
   ctx.closePath();
 }